use same type as the property in the underlying model

+++ abremir.AllMyBricks.ThirdParty.Brickset/Models/ParameterSetId.cs
@@ -2,6 +2,6 @@
 {
     public class ParameterSetId : ParameterApiKey
     {
-        public int SetID { get; set; }
+        public long SetID { get; set; }
     }
 }
